modify matching to make sure the author in a partial reference is the first author
modify matching to make sure the author in a partial reference is the first author in the full reference rather than any author. Added a script to download the sunrise and sunset times and day lengths for El Verde, changing the location and dates to get this for any location and dates.